Refactor writeGenesisBlock.
