+++ lib/blocks.js
@@ -12,6 +12,7 @@ const bedrock = require('bedrock');
 const cache = require('bedrock-redis');
 const {config} = bedrock;
 const database = require('bedrock-mongodb');
+const hasher = require('bedrock-ledger-node').consensus._hasher;
 // const logger = require('./logger');
 const {BedrockError} = bedrock.util;
 
@@ -187,6 +188,18 @@ api.write = ({ledgerNode, state, consensusResult}, callback) => {
       }, callback)],
   }, err => callback(err, err ? false : true));
 };
+
+api.writeGenesis = ({block, ledgerNode}, callback) => async.auto({
+  hashBlock: callback => hasher(block, callback),
+  writeBlock: ['hashBlock', (results, callback) => {
+    const meta = {
+      blockHash: results.hashBlock,
+      consensus: true,
+      consensusDate: Date.now()
+    };
+    ledgerNode.storage.blocks.add({block, meta}, callback);
+  }]
+}, callback);
 
 // TODO: proofs should be stored and retrieved in the same fashion as events
 function _expandConsensusProofEvents({block, ledgerNode}, callback) {
