+++ lib/cache-key.js
@@ -11,6 +11,9 @@ const database = require('bedrock-mongodb');
 const api = {};
 module.exports = api;
 
+// used to track latest blockHeight
+api.blockHeight = ledgerNodeId => `bh|${_lni(ledgerNodeId)}`;
+
 api.childless = ledgerNodeId => `cl|${_lni(ledgerNodeId)}`;
 
 api.event = ({eventHash, ledgerNodeId}) =>
@@ -35,7 +38,7 @@ api.headGeneration = ({eventHash, ledgerNodeId}) =>
   `hg|${_lni(ledgerNodeId)}|${eventHash}`;
 
 // hash portion of the voterId
-api.gossipNotification = creatorId => `gn-${_ci(creatorId)}`;
+api.gossipNotification = creatorId => `gn|${_ci(creatorId)}`;
 
 api.lastPeerHeads = ledgerNodeId => `lph|${_lni(ledgerNodeId)}`;
