+++ lib/events.js
@@ -4,6 +4,7 @@
 'use strict';
 
 const _ = require('lodash');
+const _blocks = require('./blocks');
 const _cacheKey = require('./cache-key');
 const _voters = require('./voters');
 const async = require('async');
@@ -186,8 +187,8 @@ api.add = (
         };
         callback(null, block);
       }],
-      writeBlock: ['createBlock', (results, callback) => _writeGenesisBlock({
-        block: results.createBlock, eventRecord, storage}, callback)],
+      writeBlock: ['createBlock', (results, callback) => _blocks.writeGenesis({
+        block: results.createBlock, ledgerNode}, callback)],
       // mark event has having achieved consensus
       updateEvent: ['writeBlock', (results, callback) => storage.events.update({
         eventHash: eventRecord.meta.eventHash, patch: [{
@@ -1731,18 +1732,4 @@ function _writeEvent(
     }
     callback(null, results.writeEvent);
   });
-}
-
-function _writeGenesisBlock({block, storage}, callback) {
-  async.auto({
-    hashBlock: callback => hasher(block, callback),
-    writeBlock: ['hashBlock', (results, callback) => {
-      const meta = {
-        blockHash: results.hashBlock,
-        consensus: true,
-        consensusDate: Date.now()
-      };
-      storage.blocks.add({block, meta}, callback);
-    }]
-  }, callback);
 }
